# backend/blog/src/main/java/com/itacademy/blog/data/entity/Role.java
package com.itacademy.blog.data.entity;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;

@Data
@Entity
@AllArgsConstructor
@NoArgsConstructor
@Builder
@Table(name = "role")
public class Role {
    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;

    @Enumerated(EnumType.STRING)
    private NameEnum name = NameEnum.BLOGGER;

    public enum NameEnum {
        BLOGGER,
        MODERATOR,
        ADMIN

    }
}