# backend/blog/src/main/java/com/itacademy/blog/services/DTO/RoleDTO.java
package com.itacademy.blog.services.DTO;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class RoleDTO {
    private NameEnum name;
}