# gen/api/openapi.yaml
openapi: 3.0.3
info:
  description: |
    Blog for the Home Project engineers
  license:
    name: MIT
    url: https://github.com/Home-Project-Engineering/home-project-blog/blob/main/LICENSE
  title: Home Project Blog Service
  version: 1.1.1
servers:
- url: http://localhost:8080/api/1
security:
- basicAuth: []
...
paths:
...
  /users:
    get:
      description: Uses query keys to retrieve a subset of existing users.
      operationId: getUsers
      parameters:
      - explode: true
        in: query
        name: id
        required: false
        schema:
          format: int64
          type: integer
        style: form
      - explode: true
        in: query
        name: name
        required: false
        schema:
          type: string
        style: form
      - description: |
          In order to execute *asc*, you need to specify in the search *id* or *name* parameter.
          In order to complete the *desc* sorting must be specified in the query parameter *-id*
        explode: true
        in: query
        name: sort
        required: false
        schema:
          default: -id
          enum:
          - id
          - -id
          - name
          - -name
          type: string
        style: form
      - explode: true
        in: query
        name: page_num
        required: false
        schema:
          default: 0
          type: integer
        style: form
      - explode: true
        in: query
        name: page_size
        required: false
        schema:
          default: 10
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/User'
                type: array
          description: OK
          headers:
            X-Total-Count:
              $ref: '#/components/headers/X-Total-Count'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Get users
      tags:
      - Users
      x-accepts: application/json
    post:
      description: Create a new User.
      operationId: createUser
      requestBody:
        $ref: '#/components/requestBodies/UserBody'
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Create User
      tags:
      - Users
      x-contentType: application/json
      x-accepts: application/json
  /users/{id}:
    delete:
      description: Deletes a User associated with a specified ID.
      operationId: removeUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "204":
          description: The request was succesfully processed.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Remove User
      tags:
      - Users
      x-accepts: application/json
    get:
      description: Retrieves an existing User associated with the specified ID.
      operationId: getUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Get User by ID
      tags:
      - Users
      x-accepts: application/json
    put:
      description: Updates information on a User that exists in the system by the
        specified ID.
      operationId: updateUser
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/UserBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Update User
      tags:
      - Users
      x-contentType: application/json
      x-accepts: application/json
  /users/{id}/role:
    get:
      description: Retrieves an existing User Role associated with the specified ID.
      operationId: getUserRole
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Get User Role by ID
      tags:
      - Users
      x-accepts: application/json
    put:
      description: Updates Role on a User that exists in the system by the specified
        ID.
      operationId: updateUserRole
      parameters:
      - explode: false
        in: path
        name: id
        required: true
        schema:
          format: int64
          type: integer
        style: simple
      requestBody:
        $ref: '#/components/requestBodies/RoleBody'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The payload contains an error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The specified resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: The unknown error appeard. Check your payload or contact support.
      summary: Update User Role
      tags:
      - Users
      x-contentType: application/json
      x-accepts: application/json
...
components:
  headers:
    X-Total-Count:
      description: The total results for the particular query.
      explode: false
      schema:
        format: int64
        type: number
      style: simple
...
  requestBodies:
...
    UserBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/User'
      required: true
    RoleBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Role'
      required: true
...
  schemas:
    User:
      description: A representation of User entity.
      example:
        firstName: John
        lastName: Smith
        password: passworD321
        role: ""
        name: John78
        id: 0
        email: john.smith@example.com
      properties:
        id:
          description: The ID of the User.
          format: int64
          readOnly: true
          type: integer
        name:
          description: The name of the User.
          example: John78
          minLength: 4
          type: string
        firstName:
          description: The first name of the User.
          example: John
          minLength: 1
          type: string
        lastName:
          description: The last name of the User.
          example: Smith
          minLength: 1
          type: string
        email:
          description: The email of the User.
          example: john.smith@example.com
          format: email
          type: string
        password:
          description: The password of the User. MUST contain a mix of upper and lower
            case letters, as well as digits.
          example: passworD321
          format: password
          minLength: 8
          pattern: ^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z]).{8,32}$
          type: string
        role:
          allOf:
          - $ref: '#/components/schemas/Role'
          readOnly: true
      required:
      - email
      - name
      type: object
    Role:
      description: This is the level of User access to various functions.
      example:
        name: blogger
      properties:
        name:
          default: blogger
          description: The name of the Role.
          enum:
          - blogger
          - moderator
          - admin
          type: string
      required:
      - role
      type: object
...
    Error:
      description: Representation of an Error that can appear using the application.
      properties:
        code:
          description: The code of an error that describes the Error.
          readOnly: true
          type: string
        message:
          description: The message of an error that describes the Error.
          readOnly: true
          type: string
      type: object
...
  securitySchemes:
    basicAuth:
      scheme: basic
      type: http